requirements.txt
datasets==3.3.2
fitz==0.0.1.dev2
huggingface_hub==0.29.1
mido==1.3.3
numpy==2.2.3
openai==1.64.0
pandas==2.2.3
peft==0.14.0
Pillow==11.1.0
pydub==0.25.1
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
